[level 0] Title: 두 수의 나눗셈, Time: 0.01 ms, Memory: 3.68 MB -BaekjoonHub

+++ 프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.cpp
@@ -4,6 +4,6 @@
 using namespace std;
 
 int solution(int num1, int num2) {
-    int r=num1/(float)num2*1000;
+    int r=num1/(double)num2*1000;
     return r;
 }
